Paginate blog, fix steps
- Put blog posts on separate pages
- Reworked layout.tsx's
- Fixed a bug where Petrus blocks weren't getting detected
- Enhanced the step detection in getLineStepInfo to try not to show as many useless icons like EOLine or blocks.
