+++ app/notimer/layout.tsx
@@ -1,0 +1,9 @@
+import CenteredGrid from "../../components/CenteredGrid";
+
+export default function NotimerLayout({
+  children,
+}: {
+  children: React.ReactNode;
+}) {
+  return <CenteredGrid>{children}</CenteredGrid>;
+}
