+++ app/recon/layout.tsx
@@ -1,0 +1,10 @@
+import CenteredGrid from "../../components/CenteredGrid";
+import ReconSkeleton from "../../components/recon/ReconSkeleton";
+
+export default function ReconLayout({
+  children,
+}: {
+  children: React.ReactNode;
+}) {
+  return <CenteredGrid fallback={<ReconSkeleton />}>{children}</CenteredGrid>;
+}
